+++ lib/rorvswild.rb
@@ -7,6 +7,8 @@ require "rorvswild/section"
 require "rorvswild/client"
 require "rorvswild/plugins"
 require "rorvswild/queue"
+require "rorvswild/error"
+require "rorvswild/execution"
 require "rorvswild/agent"
 
 module RorVsWild