+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Add plugin RailsError to subscribre to ActiveSupport::ErrorReporter
+
+  Any errors forwarded to `Rails.error` is sent to RorVsWild.
+  Thus you can call `Rails.error.report` instead of `RorVsWild.record_error`,
+  and `Rails.error.handle` instead of `RorVsWild.catch_error`.
+
 * Remove ASCII and Server-Timing header
 * Skip measuring empty collection renderings
 * Instrument Rails.cache via ActiveSupport notifications