+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Fix MongoDB monitoring
 * Add root section to Sidekiq tn ensure not missing any timings
 * Save local requests, jobs and errors in NSJSON instead of JSON
 * Remove ASCII and Server-Timing header